requirements.txt
django==5.0
loguru==0.7.0
Pillow==10.0.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
pandas==2.2.2
PyJWT==2.8.0
cryptography==42.0.5
django-allauth==0.61.1